Move tankerkoenig coordinator and entity to its own file
